refactor(data-layer): move models out of controllers folder
