+++ src/main/controllers/tx-history-service.js
@@ -7,9 +7,9 @@ const request = require('request');
 const async = require('async');
 const BigNumber = require('bignumber.js');
 
-const Wallet = require('./models/wallet');
-const WalletSetting = require('./models/wallet-setting');
-const TxHistory = require('./models/tx-history');
+const Wallet = require('../models/wallet');
+const WalletSetting = require('../models/wallet-setting');
+const TxHistory = require('../models/tx-history');
 
 let isSyncingMap = {};
 let syncingJobIsStarted = false;
