+++ src/main/models/guide-setting.js
@@ -7,14 +7,15 @@ const mod = (module.exports = {
 	findAll: tx => sqlUtil.select(TABLE_NAME, '*', null, tx),
 	hasAgreedToCrashReport: () => {
 		return crashReportAgreement;
-    },
-    loadCrashReportAgreement: async () => {
-        try{
-            const guideSettings = await mod.findAll();
-            crashReportAgreement = guideSettings.length > 0 ? guideSettings[0].crashReportAgreement === 1 : false;
-        } catch(e){
-            log.error(e);
-        }
+	},
+	loadCrashReportAgreement: async () => {
+		try {
+			const guideSettings = await mod.findAll();
+			crashReportAgreement =
+				guideSettings.length > 0 ? guideSettings[0].crashReportAgreement === 1 : false;
+		} catch (e) {
+			log.error(e);
+		}
 	},
 	updateById: (id, data, tx) => {
 		crashReportAgreement = data.crashReportAgreement;
