+++ src/main/models/wallet-setting.js
